[level 3] Title: 야근 지수, Time: 125.12 ms, Memory: 70.8 MB -BaekjoonHub

+++ 프로그래머스/3/12927. 야근 지수/야근 지수.java
@@ -18,7 +18,7 @@ class Solution {
         }
         
         for(int i=0;i<works.length;i++){
-            answer += (long)Math.pow(pq.poll(),2);
+            answer += Math.pow(pq.poll(),2);
         }
         
         return answer;
